Refactor ensembler
