+++ zshot/linker/linker_ensemble/__init__.py
@@ -1,0 +1,1 @@
+from zshot.linker.linker_ensemble.linker_ensemble import LinkerEnsemble  # noqa: F401
